Move background worker into its own module
